Matching flow - happy path test added, 'oldest best of five'
